Register with decorator, add more checks
