+++ packages/llm/torch_to_nnef_llm/models/handlers/openelm.py
@@ -3,9 +3,12 @@ from functools import partial
 from torch_to_nnef_llm.models.base import BaseCausal
 
 from .default import DefaultArchitectureHandler
+from .registry import register_handler
 
+
+@register_handler
 class OpenELMArchitectureHandler(DefaultArchitectureHandler):
-    """Handler for OpenELM models"""
+    """Handler for OpenELM models."""
 
     ARCH_NAMES = ("openelm",)
