+++ packages/llm/torch_to_nnef_llm/models/handlers/phi.py
@@ -1,10 +1,12 @@
 from torch_to_nnef_llm.models.base import BaseCausalWithDynCacheAndTriu
 
 from .default import DefaultArchitectureHandler
+from .registry import register_handler
 
 
+@register_handler
 class PhiArchitectureHandler(DefaultArchitectureHandler):
-    """Handler for Phi-family models"""
+    """Handler for Phi-family models."""
 
     ARCH_NAMES = ("phi",)
